Organizing release package: move ReleaseContext to its own file
